LayoutTests/pointer-lock/pointerlockchange-event-on-lock-lost.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../http/tests/resources/js-test-pre.js"></script>
<script src="../http/tests/resources/pointer-lock/pointer-lock-test-harness.js"></script>
</head>
<body>
<div>
  <div id="target1"></div>
</div>
<script>
    description("Test that webkitpointerlockchange event is dispatched when lock is lost.")
    window.jsTestIsAsync = true;
    shouldBeDefined("window.testRunner");

    targetDiv1 = document.getElementById("target1");

    todo = [
        function () {
            expectOnlyChangeEvent("Lock targetDiv1.");
            targetDiv1.webkitRequestPointerLock();
            // doNextStep called by event handler.
        },
        function () {
            shouldBe("document.webkitPointerLockElement", "targetDiv1");
            expectOnlyChangeEvent("Host initiated unlock.");
            testRunner.didLosePointerLock();
            // doNextStep called by event handler.
        },
        function () {
            shouldBe("document.webkitPointerLockElement", "null");
            doNextStep();
        },
    ];
    doNextStepWithUserGesture();
</script>
<script src="../http/tests/resources/js-test-post.js"></script>
</body>
</html>
