LayoutTests/pointer-lock/pointerlockchange-event-on-lock-lost-expected.txt

Test that webkitpointerlockchange event is dispatched when lock is lost.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.testRunner is defined.
     Lock targetDiv1.
PASS onwebkitpointerlockchange received after: Lock targetDiv1.
PASS document.webkitPointerLockElement is targetDiv1
     Host initiated unlock.
PASS onwebkitpointerlockchange received after: Host initiated unlock.
PASS document.webkitPointerLockElement is null
PASS successfullyParsed is true

TEST COMPLETE
